requirements.txt
pyaudiowpatch
pyaudiowpatch
faster-whisper
ollama
PyQt6
numpy
scipy
keyboard
google-genai
python-dotenv
fastapi
uvicorn
supabase
python-dotenv